// massive/src/types/RegisterClientInputs.ts
export interface RegisterClientInputs {
  nome: HTMLInputElement;
  cpf: HTMLInputElement;
  rg: HTMLInputElement;
  nascimento: HTMLInputElement;
  email: HTMLInputElement;
  tel: HTMLInputElement;

  cnpj: HTMLInputElement;
  ie: HTMLInputElement;
  razao: HTMLInputElement;
  fantasia: HTMLInputElement;
  abreH: HTMLInputElement;
  abreM: HTMLInputElement;
  fechaH: HTMLInputElement;
  fechaM: HTMLInputElement;

  cep: HTMLInputElement;
  endereco: HTMLInputElement;
  numero: HTMLInputElement;
  bairro: HTMLInputElement;
  cidade: HTMLInputElement;
  estado: HTMLInputElement;

  setor: HTMLSelectElement;

  diaSemana: HTMLSelectElement;
  frequencia: HTMLSelectElement;
  sequencia: HTMLInputElement;

  latitude: HTMLInputElement;
  longitude: HTMLInputElement;
}